GUI Update 1
Made changes mostly in Control and GUI. For Control, I moved some stuff around and made methods for initializing the game, mostly so that I could ensure that the GUI gets the information that it needs to display. For GUI I did mostly backend work and havent gotten to the actual display yet. Most of the calculations about what to display are done, and I just need to set up the actual display of that info now. I also still need to set up user interaction. ***There is currently no interaction with the GUI that causes the game to pause so it will loop infinitely, but there doesnt seem to be any actual errors besides that.